requirements.txt
pytorch 
torchvision 
torchaudio